提交

+++ src/sim.cpp
@@ -2490,7 +2490,7 @@ namespace madEscape
 
     // 获取每个阶段通信（ring）次数
     inline int get_comm_count_per_phase(CollectiveCommType comm_type,
-                                        CommImplementationType comm_implementation_type,int nodes_in_ring)
+                                        CommImplementationType comm_implementation_type, int nodes_in_ring)
     {
         int stream_count = 0;
         switch (comm_type)
@@ -2498,7 +2498,7 @@ namespace madEscape
         case CollectiveCommType::ALL_REDUCE:
             stream_count = 2 * (nodes_in_ring - 1);
             break;
-            case CollectiveCommType::ALL_TO_ALL:
+        case CollectiveCommType::ALL_TO_ALL:
             stream_count = ((nodes_in_ring - 1) * nodes_in_ring) / 2;
             break;
         default:
@@ -2534,6 +2534,17 @@ namespace madEscape
             msg_size = data_size;
             break;
         }
+    }
+
+    inline Direction get_comm_Ring_Direction()
+    {
+        //      if ((backend != AstraNetworkAPI::BackendType::Garnet || level > 0) && queues.size() > 1 &&
+        //     allocator >= (queues.size() / 2)) {
+        //     dir = RingTopology::Direction::Anticlockwise;
+        // } else {
+        //     dir = RingTopology::Direction::Clockwise;
+        // }
+        return Direction::Clockwise;
     }
 
     inline void processNpuNodes(Engine &ctx,
@@ -2706,153 +2717,147 @@ namespace madEscape
                 {
 
                     // NPUID*10000+FLOWID
-
-                    // 划分stream
-                    // uint64_t Sys::determine_chunk_size(uint64_t& size, ComType type) {
-                    //     uint64_t chunk_size = size / preferred_dataset_splits;
-                    //     // We want the collective size to have minimum size, otherwise, there is a
-                    //     // possibility of size overflow due to further dividing it to more
-                    //     // fine-grained messages
-                    //     if (type != ComType::All_Gather && this->total_nodes > chunk_size) {
-                    //         chunk_size = this->total_nodes;
-                    //         size = preferred_dataset_splits * chunk_size;
-                    //     }
-                    //     return chunk_size;
-                    // }
-
                     if (SYS_LOG && id.value == 0)
                     {
                         printf("processingCommTask: coll .\n");
                     }
                     // slice info, for temp
                     int npu_num = 16;
-                    int streams_num = 2;
-                    int chunk_size = node.comm_size / streams_num;
-                    int remain_size = node.comm_size % streams_num;
+                    int chunks_num = 2;
+                    int chunk_size = node.comm_size / chunks_num;
+                    int remain_size = node.comm_size % chunks_num;
 
-                    //  enum CollectiveCommType : uint64_t
-                    // {
-                    //     ALL_REDUCE = 0,
-                    //     REDUCE = 1,
-                    //     ALL_GATHER = 2,
-                    //     GATHER = 3,
-                    //     SCATTER = 4,
-                    //     BROADCAST = 5,
-                    //     ALL_TO_ALL = 6,
-                    //     REDUCE_SCATTER = 7,
-                    //     REDUCE_SCATTER_BLOCK = 8,
-                    //     BARRIER = 9
-                    // };
-
-                    for (size_t i = 0; i < chunk_size; i++)
+                    if (remain_size > 0)
                     {
-                        /* code */
+                        chunks_num++;
                     }
-                    
-                    int data_size = chunk_size;
-                    switch (node.comm_type)
+
+                    for (size_t i = 0; i < chunks_num; i++)
                     {
-                        printf("comm type: %d\n", node.comm_type);
-                    case CollectiveCommType::ALL_REDUCE:
-
-                        // 拆解成细分阶段->Reduce_Scatter+All_Gather
-                        // break;
-                    case CollectiveCommType::REDUCE:
-                        // break;
-                    case CollectiveCommType::ALL_GATHER:
-                        // break;
-                    case CollectiveCommType::GATHER:
-                        // break;
-                    case CollectiveCommType::SCATTER:
-                        // break;
-                    case CollectiveCommType::BROADCAST:
-                        // break;
-                    default:
-                    {
-                        // if (SYS_LOG && id.value == 0)
-                        // {
-                        //     printf("processingCommTask: unknown comm type .\n");
-                        // }
-
-                        //
-                        // Entity sc = ctx.makeEntity<SysConfig>();
-                        // ctx.get<CommModel>(sc).all_reduce_implementation = CommImplementationType::Ring;
-                        // ctx.get<CommModel>(sc).all_gather_implementation = CommImplementationType::Ring;
-                        // ctx.get<CommModel>(sc).reduce_scatter_implementation = CommImplementationType::Ring;
-                        // ctx.get<CommModel>(sc).all_to_all_implementation = CommImplementationType::Ring;
-                        // ctx.data().sys_config_entity = sc;
-
-                        // step1: get comm implementation
-                        CommImplementationType comm_implementation_type = ctx.get<CommModel>(ctx.data().sys_config_entity).all_reduce_implementation;
-                        switch (comm_implementation_type)
+                        int data_size_current = chunk_size;
+                        if (i == chunks_num - 1)
                         {
-                        case CommImplementationType::Ring:
-                            // step2: get logic topo
+                            data_size_current = remain_size;
+                        }
 
-                            // node info
-                            // int node_id = node.id;
-                            // int node_type = node.type;
-                            // int node_comm_type = node.comm_type;
-                            // int node_comm_size = node.comm_size;
-                            // int node_comm_src = node.comm_src;
-                            // int node_comm_dst = node.comm_dst;
-                            // int node_comm_src_npu = node.comm_src_npu;
-
-                            // topo info for temp 3d
-                            int dims[] = {4, 4, 2};
-                            int dims_len = sizeof(dims) / sizeof(dims[0]);
-                            // demotion excute info for temp
-                            bool dim_1d_enable = node.involved_dim_1;
-                            bool dim_2d_enable = node.involved_dim_2;
-                            bool dim_3d_enable = node.involved_dim_3;
-
-                            if (dim_1d_enable)
+                        switch (node.comm_type)
+                        {
+                            printf("comm type: %d\n", node.comm_type);
+                        case CollectiveCommType::ALL_REDUCE:
+                            // 拆解成细分阶段->Reduce_Scatter+All_Gather
+                            // break;
+                        case CollectiveCommType::REDUCE:
+                            // break;
+                        case CollectiveCommType::ALL_GATHER:
+                            // break;
+                        case CollectiveCommType::GATHER:
+                            // break;
+                        case CollectiveCommType::SCATTER:
+                            // break;
+                        case CollectiveCommType::BROADCAST:
+                            // break;
+                        default:
+                        {
+                            // step1: get comm implementation
+                            CommImplementationType comm_implementation_type = ctx.get<CommModel>(ctx.data().sys_config_entity).all_reduce_implementation;
+                            switch (comm_implementation_type)
                             {
-                                int total_nodes_in_ring = dims[0];
-                                int node_id = node.id;
+                            case CommImplementationType::Ring:
+                                // topo info for temp 3d
+                                int dims[] = {4, 4, 2};
+                                int dims_len = sizeof(dims) / sizeof(dims[0]);
+                                // demotion excute info for temp
+                                bool dim_1d_enable = node.involved_dim_1;
+                                bool dim_2d_enable = node.involved_dim_2;
+                                bool dim_3d_enable = node.involved_dim_3;
+
+                                // create comm task
+                                ProcessingCommTask processingCommTask = ProcessingCommTask();
+                                processingCommTask.state = TaskState::START;
+                                processingCommTask.node_id = node.id;
+                                processingCommTask.flow_count = 1;
+                                processingCommTasks.addTask(processingCommTask);
+
+                                // create comm entity
+                                Entity process_e = ctx.makeEntity<ProcessComm_E>();
+                                ctx.get<TaskFlows>(process_e) = TaskFlows();
+                                ctx.get<NpuID>(process_e).value = id.value;
+                                ctx.get<NodeID>(process_e).value = node.id;
+
+                                ctx.data().node_flows_exec_entity[id.value][node.id] = process_e;
+
+                                int flow_exec_index = 0;
+                                int flow_current_count = 0;
+
+                                int temp = 0;
                                 int offset = 1;
-                                int index = (node_id % (offset * total_nodes_in_ring)) / offset;
+                                while (temp < 3)
+                                {
+                                    bool excute = false;
+                                    int total_nodes_in_ring = dims[0];
+                                    Dimension dimension = Dimension::Local;
 
-                                // current logic topo
-                                RingTopology ring_topo(Dimension::Local, node_id, total_nodes_in_ring, index, offset);
+                                    if (temp == 0 && dim_1d_enable)
+                                    {
+                                        excute = true;
+                                    }
+                                    if (temp == 1 && dim_2d_enable)
+                                    {
+                                        total_nodes_in_ring = dims[1];
+                                        dimension = Dimension::Horizontal;
+                                        excute = true;
+                                    }
+                                    if (temp == 2 && dim_3d_enable)
+                                    {
+                                        total_nodes_in_ring = dims[2];
+                                        dimension = Dimension::Vertical;
+                                        excute = true;
+                                    }
 
-                                int flow_count = get_comm_count_per_phase(node.comm_type, comm_implementation_type,total_nodes_in_ring);
-                                printf("get_comm_count_per_phase: %d\n", flow_count);
-                                int msg_size = 0;
-                                int final_data_size = 0;
-                                get_comm_size_per_flow(node.comm_type, comm_implementation_type, data_size,total_nodes_in_ring, msg_size, final_data_size);
+                                    if (excute)
+                                    {
+                                        if (true)
+                                        {
+                                            int node_id = node.id;
+
+                                            int index = (node_id % (offset * total_nodes_in_ring)) / offset;
+                                            // current logic topo
+                                            RingTopology ring_topo(dimension, node_id, total_nodes_in_ring, index, offset);
+                                            Direction dir = get_comm_Ring_Direction();
+                                            printf("Direction: %d\n", dir);
+                                            int flow_count = get_comm_count_per_phase(node.comm_type, comm_implementation_type, total_nodes_in_ring);
+                                            printf("get_comm_count_per_phase: %d\n", flow_count);
+                                            int msg_size = 0;
+                                            int final_data_size = 0;
+                                            get_comm_size_per_flow(node.comm_type, comm_implementation_type, data_size_current, total_nodes_in_ring, msg_size, final_data_size);
+
+                                            for (size_t i = 0; i < flow_count; i++)
+                                            {
+                                                uint32_t flow_id = processingCommTasks.getTotalFlowCount();
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count] = SysFlow();
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].id = flow_id;
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].comm_size = msg_size;
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].comm_src = ring_topo.get_sender(node_id, dir);
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].comm_dst = ring_topo.get_receiver(node_id, dir);
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].state = TaskState::START;
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].is_send = true;
+                                                ctx.get<TaskFlows>(process_e).flows[flow_current_count].exec_index = flow_exec_index;
+                                                flow_current_count++;
+                                                flow_exec_index++;
+                                            }
+                                        }
+                                    }
+                                    offset *= dims[temp];
+                                    temp++;
+                                }
+
+                                break;
                             }
-
-                        break;
+                        }
                         }
                     }
-                    }
-
-                    ProcessingCommTask processingCommTask = ProcessingCommTask();
-
-                    processingCommTask.state = TaskState::START;
-                    processingCommTask.node_id = node.id;
-                    processingCommTask.flow_count = 1;
-                    processingCommTasks.addTask(processingCommTask);
-
-                    // create comm entity
-                    Entity process_e = ctx.makeEntity<ProcessComm_E>();
-                    ctx.get<TaskFlows>(process_e) = TaskFlows();
-                    ctx.get<NpuID>(process_e).value = id.value;
-                    ctx.get<NodeID>(process_e).value = node.id;
-                    uint32_t flow_id = processingCommTasks.getTotalFlowCount();
-                    ctx.get<TaskFlows>(process_e).flows[0] = SysFlow();
-                    ctx.get<TaskFlows>(process_e).flows[0].id = flow_id;
-                    ctx.get<TaskFlows>(process_e).flows[0].comm_size = 1000;
-                    ctx.get<TaskFlows>(process_e).flows[0].comm_src = 0;
-                    ctx.get<TaskFlows>(process_e).flows[0].comm_dst = 1;
-                    ctx.get<TaskFlows>(process_e).flows[0].state = TaskState::START;
-                    ctx.get<TaskFlows>(process_e).flows[0].is_send = true;
-
-                    ctx.data().node_flows_exec_entity[id.value][node.id] = process_e;
-
                     // setFlow(Engine &ctx, uint64_t comm_src, uint64_t comm_dst, uint64_t comm_size, uint32_t flow_id)
-                    setFlow(ctx, 0, 1, 1000, flow_id);
+                    // setFlow(ctx, 0, 1, 1000, flow_id);
                     break;
                 }
                 default:
@@ -2885,21 +2890,38 @@ namespace madEscape
             }
         }
 
-        SysFlow flows_finish[MAX_FLOW_NUM_PER_COMM_NODE];
-        uint32_t flow_finish_count = checkFlowFinish(ctx, npu_id.value, flows_finish);
-
-        if (flow_finish_count > 0)
+        if (!taskFlows.is_exec)
         {
-            taskFlows.updateFlows(flows_finish, flow_finish_count);
+            taskFlows.is_exec = true;
+            SysFlow flows_exec[MAX_FLOW_NUM_PER_COMM_NODE];
+            uint32_t flow_exec_count = taskFlows.getFlowsWithMinExecIndex(flows_exec);
+            if (flow_exec_count > 0)
+            {
+                for (size_t i = 0; i < flow_exec_count; i++)
+                {
+                    setFlow(ctx, flows_exec[i].comm_src, flows_exec[i].comm_dst, flows_exec[i].comm_size, flows_exec[i].id);
+                }
+            }
         }
-
-        if (taskFlows.areAllTasksDone())
+        else
         {
-            printf("node id %d -> taskFlows.areAllTasksDone\n", node_id);
 
-            ctx.get<ProcessingCommTasks>(ctx.data().chakra_nodes_entities[npu_id.value]).setFinish(node_id.value, getCurrentTime(ctx));
+            SysFlow flows_finish[MAX_FLOW_NUM_PER_COMM_NODE];
+            uint32_t flow_finish_count = checkFlowFinish(ctx, npu_id.value, flows_finish);
 
-            ctx.destroyEntity(ctx.data().node_flows_exec_entity[npu_id.value][node_id.value]);
+            if (flow_finish_count > 0)
+            {
+                taskFlows.updateFlows(flows_finish, flow_finish_count);
+            }
+
+            if (taskFlows.areAllTasksDone())
+            {
+                printf("node id %d -> taskFlows.areAllTasksDone\n", node_id);
+
+                ctx.get<ProcessingCommTasks>(ctx.data().chakra_nodes_entities[npu_id.value]).setFinish(node_id.value, getCurrentTime(ctx));
+
+                ctx.destroyEntity(ctx.data().node_flows_exec_entity[npu_id.value][node_id.value]);
+            }
         }
     }
